software/MeanFunction/src/main/java/andrewmartin/lambda/Stats.java
package andrewmartin.lambda;

import com.fasterxml.jackson.core.JsonProcessingException;
import com.fasterxml.jackson.databind.ObjectMapper;

public class Stats {
    private double mean;

    public Stats(double mean) {
        this.mean = mean;
    }

    public double getMean() {
        return mean;
    }

    public String toJson() {
        try {
            ObjectMapper objectMapper = new ObjectMapper();
            return objectMapper.writeValueAsString(this);
        } catch (JsonProcessingException e) {
            return "{}";
        }
    }
}
